+++ lib/petal_components.ex
@@ -1,7 +1,6 @@
 defmodule PetalComponents do
   defmacro __using__(_) do
     quote do
-      alias PetalComponents.Heroicons
 
       import PetalComponents.{
         Alert,